Python wheel for pastel signer -  is built

+++ lib/include/uint256.h
@@ -232,7 +232,7 @@ namespace std
             // Modify 'seed' by XORing and bit-shifting in
             // one member of after the other
             auto p = key.begin();
-            for (int i = 0; i < N; ++i)
+            for (size_t i = 0; i < N; ++i)
             {
                 hash_combine<uint64_t>(seed, *reinterpret_cast<const uint64_t*>(p));
                 p += sizeof(uint64_t);
